refactor(financeiro): extrai abas VisaoGeral/ContasReceber/ContasPagar (819→299)
Phase-5 do split do FinanceiroDashboard: as 3 abas grandes restantes saem para
componentes controlados em src/components/financeiro/dashboard/. Estado de filtro
(crFilter/cpFilter/datas) e auditTarget permanecem no pai e descem via props,
preservando os useEffect de carregamento por aba — comportamento 100% idêntico.

- VisaoGeralTab: alerts, KPIs, breakdown por empresa, aging, indicadores CFO,
  regime, inadimplentes, contas correntes (fragmento; TabsContent fica no pai).
- ContasReceberTab / ContasPagarTab: filtros, datas, totalizadores, tabela, CSV,
  drawer de histórico via onAudit.
- +9 testes (3 por aba): render, estados vazios, cliques de filtro/histórico.

tsc --noEmit limpo; lint limpo; suíte 474/475 (1 flake de carga em
useKnowledgeBaseList, verde isolado).
